25d487eb
Work on rotational normalization.

We now populate pre-rule RHSs with not only properties that have
changed, but other properties that are useful if we potentially want to
normalize inputs by rotation. (e.g. Width, Height)

This has a significant affect on rule finding, because we can no longer
assume that a pre-rule RHS contains just the keys that we want to infer,
so some of the rule finding code now removes non-changing property
values in a more just-in-time fashion.