94f9d214
If an input is divided into colored segments, and the output is the size
of the individual input segments, we check whether the output pixels can
be determined as a function of the pixels of each input segment.

In this case, the equivalent of an NOT(AND(A, B)) operator is applied.